pipReq/Version 1.1/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
altgraph==0.17.3
anyio==3.7.0
appdata==2.2.0
argcomplete==3.1.1
argsense==0.5.2
arrow==1.2.3
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
autoflake==2.2.0
Babel==2.12.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.3.0
blinker==1.6.2
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
check-requirements-txt==1.2.0
click==8.1.3
click-default-group==1.2.2
clr-loader==0.2.5
colorama==0.4.6
colored==2.2.2
commonmark==0.9.1
construct==2.10.68
cookiecutter==2.1.1
coverage==7.2.7
customtkinter==5.2.0
Cython==0.29.35
darkdetect==0.8.0
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.6
docopt==0.6.2
docutils==0.20.1
easy-install==66.0.2
Eel==0.16.0
embed-python-manager==0.2.8
engineering-notation==0.8.0
Everything-Tkinter==0.2
eyed3==0.9.7
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
flaskwebgui==1.0.6
frozenlist==1.3.3
future==0.18.3
gen-exe==0.2.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imagesize==1.4.1
incremental==22.10.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.6.0
Kivy-examples==2.2.1
Kivy-Garden==0.1.5
lazy-object-proxy==1.9.0
Listreqs==0.0.5
lk-logger==5.6.2
lk-utils==2.5.6
lxml==4.9.2
MarkupSafe==2.1.3
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
numpy==1.25.0
openapi-python-client==0.14.1
packaging==23.1
pathspec==0.11.1
pefile==2023.2.7
Pillow==9.5.0
pipreqs==0.4.13
pipx==1.2.0
platformdirs==3.8.0
proglog==0.1.10
protobuf==4.23.3
proxy-tools==0.1.0
psutil==5.9.5
pur==7.1.0
pybase62==1.0.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==1.10.9
pydub==0.25.1
pyflakes==3.0.1
pygame==2.5.0
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
pymp4==1.4.0
pypacker @ https://github.com/syegulalp/pypacker/zipball/main#sha256=21c89c07a0f2c80a4987e687a72bd4116cba568b4263f95a00c02a304077d7e0
pyparsing==3.1.0
pypiwin32==223
pyportable-crypto==1.1.0
pyportable-installer==4.4.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
pyqtdeploy==3.3.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pythonnet==3.0.1
pytube==15.0.0
pytz==2023.3
pyvan @ file:///C:/Users/PUDDD/Downloads/Compressed/pyvan-master/pyvan-master
pywebview==4.2.2
pywidevine==1.6.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
redis==4.6.0
remi==2022.7.27
requests==2.31.0
requirements-detector==1.2.2
rich==13.4.2
semver==3.0.1
sentry-sdk==1.26.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.0.1
sphinx-autodoc-typehints==1.23.3
sphinx-issues==3.0.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-towncrier==0.3.1a3
spoti-dl==0.9.2
spotify-aac-downloader==1.3
spotify-dl==8.8.1
spotipy==2.23.0
sv-ttk==2.5.1
text-unidecode==1.3
tk-tools==0.16.0
tkinter-page==0.0.5
TKinterModernThemes==1.9.3
to-requirements.txt==1.3.0
toml==0.10.2
tomli==2.0.1
towncrier==23.6.0
tqdm==4.65.0
ttkbootstrap==1.10.1
ttkthemes==3.2.2
ttkwidgets==0.13.0
typer==0.9.0
typing_extensions==4.7.0
Unidecode==1.3.6
urllib3==2.0.3
userpath==1.8.0
virtualenv==20.23.1
websockets==11.0.3
Werkzeug==2.3.6
whichcraft==0.6.1
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
youtube-search==2.1.2
yt-dlp==2023.6.22
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
